# museum-service/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  contact:
    email: team@redocly.com
    url: https://redocly.com/docs/cli/
  description: "Imaginary, but delightful Museum API for interacting with museum services\
    \ and information. Built with love by Redocly."
  license:
    name: MIT
    url: https://opensource.org/license/mit/
  termsOfService: https://redocly.com/subscription-agreement/
  title: Redocly Museum API
  version: 1.1.0
servers:
- url: http://localhost:8080/v1
security:
- MuseumPlaceholderAuth: []
...
paths:
...
  /special-events:
    get:
      description: Return a list of upcoming special events at the museum.
      operationId: listSpecialEvents
      parameters:
      - $ref: '#/components/parameters/StartDate'
      - $ref: '#/components/parameters/EndDate'
      - $ref: '#/components/parameters/PaginationPage'
      - $ref: '#/components/parameters/PaginationLimit'
      responses:
        "200":
          content:
            application/json:
              examples:
                default_example:
                  $ref: '#/components/examples/ListSpecialEventsResponseExample'
              schema:
                $ref: '#/components/schemas/ListSpecialEventsResponse'
          description: Success.
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
      summary: List special events
      tags:
      - Events
      x-accepts: "application/json,application/problem+json"
      x-tags:
      - tag: Events
    post:
      description: Creates a new special event for the museum.
      operationId: createSpecialEvent
      requestBody:
        content:
          application/json:
            examples:
              default_example:
                $ref: '#/components/examples/CreateSpecialEventRequestExample'
            schema:
              $ref: '#/components/schemas/CreateSpecialEventRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              examples:
                default_example:
                  $ref: '#/components/examples/CreateSpecialEventResponseExample'
              schema:
                $ref: '#/components/schemas/SpecialEventResponse'
          description: Success.
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
      summary: Create special events
      tags:
      - Events
      x-content-type: application/json
      x-accepts: "application/json,application/problem+json"
      x-tags:
      - tag: Events
  /special-events/{eventId}:
    delete:
      description: Delete a special event from the collection. Allows museum to cancel
        planned events.
      operationId: deleteSpecialEvent
      parameters:
      - $ref: '#/components/parameters/EventId'
      responses:
        "204":
          description: Success - no content.
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
      summary: Delete special event
      tags:
      - Events
      x-accepts: application/problem+json
      x-tags:
      - tag: Events
    get:
      description: Get details about a special event.
      operationId: getSpecialEvent
      parameters:
      - $ref: '#/components/parameters/EventId'
      responses:
        "200":
          content:
            application/json:
              examples:
                default_example:
                  $ref: '#/components/examples/GetSpecialEventResponseExample'
              schema:
                $ref: '#/components/schemas/SpecialEventResponse'
          description: Success.
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
      summary: Get special event
      tags:
      - Events
      x-accepts: "application/json,application/problem+json"
      x-tags:
      - tag: Events
    patch:
      description: Update the details of a special event.
      operationId: updateSpecialEvent
      parameters:
      - $ref: '#/components/parameters/EventId'
      requestBody:
        content:
          application/json:
            examples:
              default_example:
                $ref: '#/components/examples/UpdateSpecialEventRequestExample'
            schema:
              $ref: '#/components/schemas/UpdateSpecialEventRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              examples:
                default_example:
                  $ref: '#/components/examples/UpdateSpecialEventResponseExample'
              schema:
                $ref: '#/components/schemas/SpecialEventResponse'
          description: Success.
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
      summary: Update special event
      tags:
      - Events
      x-content-type: application/json
      x-accepts: "application/json,application/problem+json"
      x-tags:
      - tag: Events
...
components:
  examples:
...
    CreateSpecialEventRequestExample:
      summary: Create special event
      value:
        name: Mermaid Treasure Identification and Analysis
        location: Under the seaaa 🦀 🎶 🌊.
        eventDescription: "Join us as we review and classify a rare collection of\
          \ 20 thingamabobs, gadgets, gizmos, whoosits, and whatsits, kindly donated\
          \ by Ariel."
        dates:
        - 2023-09-05
        - 2023-09-08
        price: 0
    CreateSpecialEventResponseExample:
      summary: Special event created
      value:
        eventId: dad4bce8-f5cb-4078-a211-995864315e39
        name: Mermaid Treasure Identification and Analysis
        location: Under the seaaa 🦀 🎶 🌊.
        eventDescription: "Join us as we review and classify a rare collection of\
          \ 20 thingamabobs, gadgets, gizmos, whoosits, and whatsits, kindly donated\
          \ by Ariel."
        dates:
        - 2023-09-05
        - 2023-09-08
        price: 30
    GetSpecialEventResponseExample:
      summary: Get special event
      value:
        eventId: 6744a0da-4121-49cd-8479-f8cc20526495
        name: Time Traveler Tea Party
        location: Temporal Tearoom
        eventDescription: Sip tea with important historical figures.
        dates:
        - 2023-11-18
        - 2023-11-25
        - 2023-12-02
        price: 60
    ListSpecialEventsResponseExample:
      summary: List of special events
      value:
      - eventId: f3e0e76e-e4a8-466e-ab9c-ae36c15b8e97
        name: Sasquatch Ballet
        location: Seattle... probably
        eventDescription: "They're big, they're hairy, but they're also graceful.\
          \ Come learn how the biggest feet can have the lightest touch."
        dates:
        - 2023-12-15
        - 2023-12-22
        price: 40
      - eventId: 2f14374a-9c65-4ee5-94b7-fba66d893483
        name: Solar Telescope Demonstration
        location: Far from the sun.
        eventDescription: Look at the sun without going blind!
        dates:
        - 2023-09-07
        - 2023-09-14
        price: 50
      - eventId: 6aaa61ba-b2aa-4868-b803-603dbbf7bfdb
        name: Cook like a Caveman
        location: Fire Pit on East side
        eventDescription: Learn to cook on an open flame.
        dates:
        - 2023-11-10
        - 2023-11-17
        - 2023-11-24
        price: 5
      - eventId: 602b75e1-5696-4ab8-8c7a-f9e13580f910
        name: Underwater Basket Weaving
        location: Rec Center Pool next door.
        eventDescription: Learn to weave baskets underwater.
        dates:
        - 2023-09-12
        - 2023-09-15
        price: 15
      - eventId: dad4bce8-f5cb-4078-a211-995864315e39
        name: Mermaid Treasure Identification and Analysis
        location: Room Sea-12
        eventDescription: "Join us as we review and classify a rare collection of\
          \ 20 thingamabobs, gadgets, gizmos, whoosits, and whatsits — kindly donated\
          \ by Ariel."
        dates:
        - 2023-09-05
        - 2023-09-08
        price: 30
      - eventId: 6744a0da-4121-49cd-8479-f8cc20526495
        name: Time Traveler Tea Party
        location: Temporal Tearoom
        eventDescription: Sip tea with important historical figures.
        dates:
        - 2023-11-18
        - 2023-11-25
        - 2023-12-02
        price: 60
      - eventId: 3be6453c-03eb-4357-ae5a-984a0e574a54
        name: Pirate Coding Workshop
        location: Computer Room
        eventDescription: Captain Blackbeard shares his love of the C...language.
          And possibly Arrrrr (R lang).
        dates:
        - 2023-10-29
        - 2023-10-30
        - 2023-10-31
        price: 45
      - eventId: 9d90d29a-2af5-4206-97d9-9ea9ceadcb78
        name: Llama Street Art Through the Ages
        location: Auditorium
        eventDescription: Llama street art?! Alpaca my bags -- let's go!
        dates:
        - 2023-10-29
        - 2023-10-30
        - 2023-10-31
        price: 45
      - eventId: a3c7b2c4-b5fb-4ef7-9322-00a919864957
        name: The Great Parrot Debate
        location: Outdoor Amphitheatre
        eventDescription: See leading parrot minds discuss important geopolitical
          issues.
        dates:
        - 2023-11-03
        - 2023-11-10
        price: 35
      - eventId: b92d46b7-4c5d-422b-87a5-287767e26f29
        name: Eat a Bunch of Corn
        location: Cafeteria
        eventDescription: We accidentally bought too much corn. Please come eat it.
        dates:
        - 2023-11-10
        - 2023-11-17
        - 2023-11-24
        price: 5
    UpdateSpecialEventRequestExample:
      summary: Update special event request
      value:
        location: On the beach.
        price: 15
    UpdateSpecialEventResponseExample:
      summary: Update special event
      value:
        eventId: dad4bce8-f5cb-4078-a211-995864315e39
        name: Mermaid Treasure Identification and Analysis
        location: On the beach.
        eventDescription: "Join us as we review and classify a rare collection of\
          \ 20 thingamabobs, gadgets, gizmos, whoosits, and whatsits, kindly donated\
          \ by Ariel."
        dates:
        - 2023-09-05
        - 2023-09-08
        price: 15
...
  parameters:
    PaginationPage:
      description: Page number to retrieve.
      explode: true
      in: query
      name: page
      required: false
      schema:
        default: 1
        example: 2
        type: integer
      style: form
    PaginationLimit:
      description: Number of days per page.
      explode: true
      in: query
      name: limit
      required: false
      schema:
        default: 10
        example: 15
        maximum: 30
        type: integer
      style: form
    EventId:
      description: Identifier for a special event.
      explode: false
      in: path
      name: eventId
      required: true
      schema:
        example: dad4bce8-f5cb-4078-a211-995864315e39
        format: uuid
        type: string
      style: simple
    StartDate:
      description: Starting date to retrieve future operating hours from. Defaults
        to today's date.
      explode: true
      in: query
      name: startDate
      required: false
      schema:
        example: 2023-02-23
        format: date
        type: string
      style: form
    EndDate:
      description: End of a date range to retrieve special events for. Defaults to
        7 days after `startDate`.
      explode: true
      in: query
      name: endDate
      required: false
      schema:
        example: 2023-04-18
        format: date
        type: string
      style: form
...
  responses:
    BadRequest:
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
      description: Bad request.
    NotFound:
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
      description: Not found.
    Unauthorized:
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
      description: Unauthorized.
  schemas:
...
    Date:
      example: 2023-10-29
      format: date
      type: string
...
    CreateSpecialEventRequest:
      description: Request payload for creating new special events at the museum.
      example:
        price: 25
        name: Pirate Coding Workshop
        eventDescription: Captain Blackbeard shares his love of the C...language.
          And possibly Arrrrr (R lang).
        location: Computer Room
        dates:
        - 2023-10-29
        - 2023-10-29
        info: info
      properties:
        name:
          description: Name of the special event.
          example: Pirate Coding Workshop
          type: string
        location:
          description: Location where the special event is held.
          example: Computer Room
          type: string
        eventDescription:
          description: Description of the special event.
          example: Captain Blackbeard shares his love of the C...language. And possibly
            Arrrrr (R lang).
          type: string
        dates:
          description: List of planned dates for the special event.
          items:
            $ref: '#/components/schemas/Date'
          type: array
        price:
          description: Price of a ticket for the special event.
          example: 25
          format: float
          type: number
        info:
          description: Additional information about the event.
          type: string
      required:
      - dates
      - eventDescription
      - info
      - location
      - name
      - price
    UpdateSpecialEventRequest:
      description: Request payload for updating an existing special event. Only included
        fields are updated in the event.
      example:
        price: 25
        name: Pirate Coding Workshop
        eventDescription: Captain Blackbeard shares his love of the C...language.
          And possibly Arrrrr (R lang).
        location: Computer Room
        dates:
        - 2023-10-29
        - 2023-10-29
      properties:
        name:
          description: Name of the special event.
          example: Pirate Coding Workshop
          type: string
        location:
          description: Location where the special event is held.
          example: Computer Room
          type: string
        eventDescription:
          description: Description of the special event.
          example: Captain Blackbeard shares his love of the C...language. And possibly
            Arrrrr (R lang).
          type: string
        dates:
          description: List of planned dates for the special event.
          items:
            $ref: '#/components/schemas/Date'
          type: array
        price:
          description: Price of a ticket for the special event.
          example: 25
          format: float
          type: number
    ListSpecialEventsResponse:
      description: List of upcoming special events.
      items:
        $ref: '#/components/schemas/SpecialEventResponse'
      type: array
    SpecialEventResponse:
      description: Information about a special event.
      example:
        eventId: 3be6453c-03eb-4357-ae5a-984a0e574a54
        price: 25
        name: Pirate Coding Workshop
        eventDescription: Captain Blackbeard shares his love of the C...language.
          And possibly Arrrrr (R lang).
        location: Computer Room
        dates:
        - 2023-10-29
        - 2023-10-29
      properties:
        eventId:
          description: Identifier for a special event.
          example: 3be6453c-03eb-4357-ae5a-984a0e574a54
          format: uuid
          type: string
        name:
          description: Name of the special event.
          example: Pirate Coding Workshop
          type: string
        location:
          description: Location where the special event is held.
          example: Computer Room
          type: string
        eventDescription:
          description: Description of the special event.
          example: Captain Blackbeard shares his love of the C...language. And possibly
            Arrrrr (R lang).
          type: string
        dates:
          description: List of planned dates for the special event.
          items:
            $ref: '#/components/schemas/Date'
          type: array
        price:
          description: Price of a ticket for the special event.
          example: 25
          format: float
          type: number
      required:
      - dates
      - eventDescription
      - eventId
      - location
      - name
      - price
    Error:
      properties:
        type:
          example: object
          type: string
        title:
          example: Validation failed
          type: string
  securitySchemes:
    MuseumPlaceholderAuth:
      scheme: basic
      type: http